requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
google-genai